Add `TARGET` param.
Previously, there was no way of having a particular stage
targetted when using the `build` script.

Being able to do so is particularly important for cases when
the default final stage does not need an intermediary testing
stage, meaning that such stage doesn't even get executed.

Now, with `--target`, one is able to have a task that explicitly
targets testing, leaving another task doing the default (or reaching
a `release` target).

For example:

```
FROM alpine AS build

	RUN echo "build" > /build

FROM alpine AS test

	RUN echo "test" > /test

FROM alpine AS release

	COPY --from=build /build /build
```

Without `--target`:

`build` and `release` are run.

With `--target=test`:

`test` is run.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>