add target_file; address `nounset` problem with TARGET
Instead of relying on one-liner for doing the whole conditional logic
around `TARGET` being set, here we adhere more to the pattern that is
already in place.

In light of https://github.com/concourse/builder-task/pull/21, this
commit also introduces `TARGET_FILE` so that one can specify the target
via a file in the fs (just like we do for other arguments).

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>