use 'docker' exporter for now
this really simplifies the unpacking implementation since it doesn't
have to guess which image to unpack - it can just use the manifest.json
file.

once a new buildkit is out we can go back to type=oci, which will also
produce the manifest.json file.

technically this is already how things work since the images are built
against moby/buildkit:master (hence the confusion in
14cf93fbfff3a533d75e9b13aa2a5ca2d8cb427c) - but the tests are still
dependent on v0.6.1 so we'll just set it to the 'docker' exporter until
the next release is out.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>