remove tag/repository config, create digest file
a repository name and tag aren't necessary for building an image, and
baking them into the image kind of feels like a legacy practice. the OCI
format has no place for the repository to be stored.

in order to continue supporting 'docker load' and identifying the loaded
image, we can instead just write the image's digest to a separate file
in the output. this digest can then be tagged to give it a name:

    docker load -i image.tar
    docker tag $(cat digest) my-name

this feels like a much cleaner split: repository/tag naming is a concern
of pushing to a registry, not of image building, so it should be limited
to the registry-pushing resource (i.e. registry-image) instead of having
to configure it in both.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>