fix inconsistent layer compression w/ image args
previously, unpacking an image built with image_args could fail if (i
think!) the FROM image's layers were uncompressed. the added layers
would be compressed, and this would cause go-containerregistry to assume
all of them were, leading to an 'unexpected EOF' (or 'expected tar
header') error when trying to unpack the compressed layers.

i think this happened because I couldn't properly return a mediatype
from the registry indicating whether the layer blob was compressed or
decompressed, because GetBlob didn't return that information, so maybe
buildkit just trusted that it was compressed...?

just a guess though. this was kind of tricky to debug.

in any case, this is fixed by dropping github.com/containers/image and
using go-containerregistry directly for the registry implementation.
should have just used it directly in the first place tbh to avoid the
dependency. much tidier now!

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>