change cache export from min to max
With the default min setting only the cache of the final image is exported resulting in poor cache performance for most multi-stage dockerfiles as the heavy work is usually happening in intermediate images.
This will lead to a bigger cache size but improves build time as more parts of multi-stage builds are cached.

Fixes #26

Signed-off-by: Fabian Ruff <fabian.ruff@sap.com>