load OCI images
Enable the oci-build task to load OCI images from a given path. The
original tarball loading still works, but now the task will detect if
the given path is a directory or a file.

If it's a directory then it tries loading the files at the path as an
OCI image and adds it to the image registry server that we spin up.

Signed-off-by: Taylor Silva <dev@taydev.net>