Set OUTPUT_OCI to true if IMAGE_PLATFORM contains multiple platforms
If a user didn't do this then they would get the error from https://github.com/concourse/oci-build-task/issues/85

This PR improves the UX of this task by automatically setting OUTPUT_OCI
to true for the user. The only gotcha that remains is that this changes
the output format from a tarball to an OCI Image Layout directory. This
is now documented in the README though.

Signed-off-by: Taylor Silva <dev@taydev.net>